+++ src/helpers/otpGenerator.helper.ts
@@ -1,0 +1,5 @@
+const generateOTP = () => {
+    return (Math.floor(1000 + Math.random() * 9000)).toString();
+};
+
+export default generateOTP;
