+++ src/libs/nodemailer.lib.ts
@@ -1,0 +1,13 @@
+import { createTransport } from "nodemailer";
+
+const transporter = createTransport({
+    host: process.env.NODEMAILER_HOST,
+    port: 587,
+    secure: false,
+    auth: {
+        user: process.env.NODEMAILER_EMAIL,
+        pass: process.env.NODEMAILER_PASSWORD,
+    },
+});
+
+export default transporter;
