+++ src/libs/prisma.lib.ts
@@ -1,0 +1,9 @@
+import { PrismaClient } from "@prisma/client";
+
+const prisma = new PrismaClient();
+
+const getPrismaClient = () => {
+    return prisma;
+};
+
+export default getPrismaClient;
