--- src/libs/prisma.ts
@@ -1,9 +1,0 @@
-import { PrismaClient } from "@prisma/client";
-
-const prisma = new PrismaClient();
-
-const getPrismaClient = () => {
-    return prisma;
-};
-
-export default getPrismaClient;
