+++ src/libs/rabbitmq.lib.ts
@@ -1,0 +1,9 @@
+import { connect } from "amqplib";
+
+const connectRabbitMQ = async () => {
+    const connection = await connect(process.env.RABBITMQ_URL as string);
+    const channel = await connection.createChannel();
+    return { connection, channel };
+};
+
+export default connectRabbitMQ;
