+++ src/middlewares/auth.middleware.ts
@@ -2,7 +2,7 @@ import { Context } from "hono";
 import { getSignedCookie } from "hono/cookie";
 
 import ApiError from "../utils/ApiError";
-import getPrismaClient from "../libs/prisma";
+import getPrismaClient from "../libs/prisma.lib";
 import { decodeToken } from "../helpers/jwtToken.helper";
 import errorMessage from "../helpers/errorMessage.helper";
